// frontend/src/client/types.gen.ts
export type IconsObject = '/src/object' | '/src/person' | '/src/bike' | '/src/scooter' | '/src/bus' | '/src/car' | '/src/tractor' | '/src/harvester' | '/src/airplane';
// ...
export type TrackingObjectPublic = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Description
     */
    description: string;
    icon_object?: IconsObject;
    /**
     * Signal
     */
    signal: number;
    /**
     * Position Lat
     */
    position_lat: number | string | null;
    /**
     * Position Lon
     */
    position_lon: number | string | null;
};